+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased][unreleased]
+### Feature
+* Added transformer feature before putting into cache (#95)
 
 ## [3.0.0-rc.3][2016-08-20]
 ### Bugfix